factor out cyclic import
