+++ src/rpc/responses.rs
@@ -59,19 +59,17 @@ use solders_macros::{
     common_methods, common_methods_rpc_resp, enum_original_mapping, richcmp_eq_only, EnumIntoPy,
 };
 use solders_primitives::{pubkey::Pubkey, signature::Signature};
-use solders_traits::{
-    py_from_bytes_general_via_bincode, pybytes_general_via_bincode, to_py_err, PyBytesBincode,
-    PyFromBytesBincode, RichcmpEqualityOnly,
-};
+use solders_traits::{to_py_err, PyBytesBincode, PyFromBytesBincode, RichcmpEqualityOnly};
 use solders_transaction_status::{
     tmp_transaction_status::{
         TransactionConfirmationStatus as TransactionConfirmationStatusOriginal,
-        TransactionStatus as TransactionStatusOriginal, UiTransactionReturnData,
+        TransactionStatus as TransactionStatusOriginal,
     },
     EncodedConfirmedTransactionWithStatusMeta, TransactionConfirmationStatus, TransactionErrorType,
-    TransactionReturnData, TransactionStatus, UiConfirmedBlock,
+    TransactionStatus, UiConfirmedBlock,
 };
 
+use super::common::{response_data_boilerplate, RpcSimulateTransactionResult};
 use super::errors::{
     BlockCleanedUpMessage, BlockNotAvailableMessage, BlockStatusNotAvailableYetMessage,
     InternalErrorMessage, InvalidParamsMessage, InvalidRequestMessage,
@@ -141,20 +139,6 @@ macro_rules! resp_traits {
             }
         }
         impl<'a> CommonMethodsRpcResp<'a> for $name {}
-    };
-}
-
-macro_rules! response_data_boilerplate {
-    ($name:ident) => {
-        impl RichcmpEqualityOnly for $name {}
-        impl Display for $name {
-            fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-                write!(f, "{:?}", self)
-            }
-        }
-        pybytes_general_via_bincode!($name);
-        py_from_bytes_general_via_bincode!($name);
-        solders_traits::common_methods_default!($name);
     };
 }
 
@@ -1226,56 +1210,6 @@ contextful_resp_eq!(GetFeeForMessageResp, Option<u64>);
 contextless_resp_eq!(GetFirstAvailableBlockResp, u64);
 contextless_resp_eq!(GetGenesisHashResp, SolderHash, "DisplayFromStr");
 contextless_resp_eq!(GetHealthResp, String, clone);
-
-// the one in solana_client doesn't derive Eq
-// TODO: latest does
-#[serde_as]
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq)]
-#[serde(rename_all = "camelCase")]
-#[pyclass(module = "solders.rpc.responses", subclass)]
-pub struct RpcSimulateTransactionResult {
-    #[serde(default)]
-    #[pyo3(get)]
-    pub err: Option<TransactionErrorType>,
-    #[serde(default)]
-    #[pyo3(get)]
-    pub logs: Option<Vec<String>>,
-    #[serde_as(as = "Option<Vec<Option<TryFromInto<UiAccount>>>>")]
-    #[serde(default)]
-    #[pyo3(get)]
-    pub accounts: Option<Vec<Option<Account>>>,
-    #[serde(default)]
-    #[pyo3(get)]
-    pub units_consumed: Option<u64>,
-    #[serde_as(as = "Option<FromInto<UiTransactionReturnData>>")]
-    #[serde(default)]
-    #[pyo3(get)]
-    pub return_data: Option<TransactionReturnData>,
-}
-
-response_data_boilerplate!(RpcSimulateTransactionResult);
-
-#[richcmp_eq_only]
-#[common_methods]
-#[pymethods]
-impl RpcSimulateTransactionResult {
-    #[new]
-    pub fn new(
-        err: Option<TransactionErrorType>,
-        logs: Option<Vec<String>>,
-        accounts: Option<Vec<Option<Account>>>,
-        units_consumed: Option<u64>,
-        return_data: Option<TransactionReturnData>,
-    ) -> Self {
-        Self {
-            err,
-            logs,
-            accounts,
-            units_consumed,
-            return_data,
-        }
-    }
-}
 
 #[derive(Debug, PartialEq, Eq, Serialize, Deserialize, Clone, From, Into)]
 #[pyclass(module = "solders.rpc.responses", subclass)]
